+++ src/components/form/select/SelectFieldMulti.tsx
@@ -8,8 +8,9 @@ import SelectFieldMultiValue from "./SelectFieldMultiValue";
 import SelectFieldOptions from "./SelectFieldOptions";
 import fuzzyFilter from "./fuzzyFilter";
 
-type SelectFieldMultiProps = FormState & {
+type SelectFieldMultiProps = Omit<FormState, "disabled"> & {
   creatable: boolean;
+  disabled?: boolean;
   inputRef: React.RefObject<HTMLInputElement>;
   name: string;
   onChange?: (value: null | SelectValue[]) => void;
@@ -180,8 +181,8 @@ class SelectFieldMulti extends React.Component<SelectFieldMultiProps, SelectFiel
   // we're following the rules for it but it can't figure that out
   render() {
     const {
-      creatable, inputRef, name, onError, options, placeholder, required,
-      selectRef, submitted, validator
+      creatable, disabled, inputRef, name, onError, options, placeholder,
+      required, selectRef, submitted, validator
     } = this.props;
 
     const { display, filteredOptions, open, touched } = this.state;
@@ -192,6 +193,7 @@ class SelectFieldMulti extends React.Component<SelectFieldMultiProps, SelectFiel
       <>
         <div ref={selectRef} className="chq-ffd--sl">
           <SelectFieldMultiValue
+            disabled={disabled}
             display={display}
             inputRef={inputRef}
             name={name}